+++ ext_emconf.php
@@ -9,7 +9,7 @@ $EM_CONF[$_EXTKEY] = [
     'version' => '3.0.0-dev',
     'constraints' => [
         'depends' => [
-            'typo3' => '11.5.0-13.4.99',
+            'typo3' => '12.4.0-13.4.99',
             'php' => '8.1.0-0.0.0',
         ],
         'conflicts' => [],